requirements.txt
fastapi
uvicorn[standard]
pydantic-settings
sqlalchemy[asyncio]
aiosqlite
passlib[bcrypt]
python-jose[cryptography]